Rearraged and renamed files, folders. New names for several functions. Added possibility to enter formulas only into number-cells
