+++ src/plugins/math/pureMath/historyFormulas.tsx
